# swagger.yaml
openapi: 3.0.0
info:
  title: GOSHOP-ECOMMERCE-API
  contact: {}
  version: '1.0'
servers:
- url: https://goshop-vercel.vercel.app/api/v1
  variables: {}
paths:
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
      - User
      summary: Get Single User
      operationId: GetSingleUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /reviews:
    get:
      tags:
      - Reviews
      summary: Get All Reviews
      operationId: GetAllReviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
      - Reviews
      summary: Create Review
      operationId: CreateReview
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateReviewRequest'
              - example:
                  title: Good Product
                  rating: 4
                  comment: This is just what I need. thanks you for the best customer support and fast shipping
                  user: 64363fd7358253c2d0b03a69
                  product: 643b1d67896a6e693e7c8823
            example:
              title: Good Product
              rating: 4
              comment: This is just what I need. thanks you for the best customer support and fast shipping
              user: 64363fd7358253c2d0b03a69
              product: 643b1d67896a6e693e7c8823
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /reviews/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The review ID
    get:
      tags:
      - Reviews
      summary: Get Single Review
      operationId: GetSingleReview
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    patch:
      tags:
      - Reviews
      summary: Update Review
      operationId: UpdateReview
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateReviewRequest'
              - example:
                  title: Good Product
                  rating: 5
                  comment: This is just what I need. thanks you for the best customer support and fast shipping
                  user: 64363fd7358253c2d0b03a69
                  product: 643b1d67896a6e693e7c8823
            example:
              title: Good Product
              rating: 5
              comment: This is just what I need. thanks you for the best customer support and fast shipping
              user: 64363fd7358253c2d0b03a69
              product: 643b1d67896a6e693e7c8823
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  
    delete:
      tags:
      - Reviews
      summary: Delete Review
      operationId: DeleteReview
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /reviews/getSingleProductReview/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The review ID
    get:
      tags:
      - Reviews
      summary: Get Single Product Reviews
      operationId: GetSingleProductReviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    UpdateReviewRequest:
      title: UpdateReviewRequest
      required:
      - title
      - rating
      - comment
      - user
      - product
      type: object
      properties:
        title:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        user:
          type: string
        product:
          type: string
      example:
        title: Good Product
        rating: 5
        comment: This is just what I need. thanks you for the best customer support and fast shipping
        user: 64363fd7358253c2d0b03a69
        product: 643b1d67896a6e693e7c8823
    CreateReviewRequest:
      title: CreateReviewRequest
      required:
      - title
      - rating
      - comment
      - user
      - product
      type: object
      properties:
        title:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        user:
          type: string
        product:
          type: string
      example:
        title: Good Product
        rating: 4
        comment: This is just what I need. thanks you for the best customer support and fast shipping
        user: 64363fd7358253c2d0b03a69
        product: 643b1d67896a6e693e7c8823
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...